Adding some 64-bit support

+++ rcbot/entprops.cpp
@@ -33,6 +33,12 @@
 #pragma pop_macro("clamp")
 
 #include "entprops.h"
+
+#include <am-string.h>
+#include <IGameConfigs.h>
+#include <IGameHelpers.h>
+#include <IPlayerHelpers.h>
+
 #include "bot_sm_ext.h"
 #include "logging.h"
 #include "helper.h"
@@ -416,7 +422,8 @@ int CBotEntProp::GetEntProp(const int entity, const PropType proptype, const cha
 
 		if (td->fieldType == FIELD_CUSTOM && (td->flags & FTYPEDESC_OUTPUT) == FTYPEDESC_OUTPUT)
 		{
-			const variant_t* pVariant = reinterpret_cast<variant_t*>(reinterpret_cast<intptr_t>(pEntity) + offset);
+			const variant_t* pVariant = reinterpret_cast<variant_t*>(reinterpret_cast<intptr_t>(pEntity) + static_cast<intptr_t>(offset));
+
 			if ((bit_count = MatchTypeDescAsInteger(pVariant->fieldType, 0)) == 0)
 			{
 				logger->Log(LogLevel::ERROR, "Variant value for %s is not an integer (%d)", prop, pVariant->fieldType);
@@ -466,25 +473,27 @@ int CBotEntProp::GetEntProp(const int entity, const PropType proptype, const cha
 
 	if (bit_count >= 17)
 	{
-		return *reinterpret_cast<int32_t*>(reinterpret_cast<uint8_t*>(pEntity) + offset);
+		return *reinterpret_cast<int32_t*>(reinterpret_cast<uint8_t*>(pEntity) + static_cast<size_t>(offset));
 	}
+
 	if (bit_count >= 9)
 	{
 		if (is_unsigned)
 		{
-			return *reinterpret_cast<uint16_t*>(reinterpret_cast<uint8_t*>(pEntity) + offset);
+			return *reinterpret_cast<uint16_t*>(reinterpret_cast<uint8_t*>(pEntity) + static_cast<size_t>(offset));
 		}
-		return *reinterpret_cast<int16_t*>(reinterpret_cast<uint8_t*>(pEntity) + offset);
+		return *reinterpret_cast<int16_t*>(reinterpret_cast<uint8_t*>(pEntity) + static_cast<size_t>(offset));
 	}
+
 	if (bit_count >= 2)
 	{
 		if (is_unsigned)
 		{
-			return *(reinterpret_cast<uint8_t*>(pEntity) + offset);
+			return *(reinterpret_cast<uint8_t*>(pEntity) + static_cast<size_t>(offset));
 		}
-		return *reinterpret_cast<int8_t*>(reinterpret_cast<uint8_t*>(pEntity) + offset);
+		return *reinterpret_cast<int8_t*>(reinterpret_cast<uint8_t*>(pEntity) + static_cast<size_t>(offset));
 	}
-	return reinterpret_cast<bool*>(reinterpret_cast<uint8_t*>(pEntity) + offset) ? 1 : 0;
+	return reinterpret_cast<bool*>(reinterpret_cast<uint8_t*>(pEntity) + static_cast<size_t>(offset)) ? 1 : 0;
 }
 
 /// @brief Retrieves an integer pointer in an entity's property.
@@ -543,7 +552,7 @@ int *CBotEntProp::GetEntPropPointer(const int entity, const PropType proptype, c
 
 		if (td->fieldType == FIELD_CUSTOM && (td->flags & FTYPEDESC_OUTPUT) == FTYPEDESC_OUTPUT)
 		{
-			const variant_t* pVariant = reinterpret_cast<variant_t*>(reinterpret_cast<intptr_t>(pEntity) + offset);
+			const variant_t* pVariant = reinterpret_cast<variant_t*>(reinterpret_cast<intptr_t>(pEntity) + static_cast<size_t>(offset));
 			if ((bit_count = MatchTypeDescAsInteger(pVariant->fieldType, 0)) == 0)
 			{
 				logger->Log(LogLevel::ERROR, "Variant value for %s is not an integer (%d)", prop, pVariant->fieldType);
