+++ utils/RCBot2_meta/bot.h
@@ -1098,7 +1098,12 @@ public:
 	static void runPlayerMoveAll ();
 
 	static CBot* get(const size_t iIndex) { return m_Bots[iIndex]; }
-	static CBot *get (const edict_t *pPlayer ) { return m_Bots[slotOfEdict(pPlayer)]; }
+
+	static CBot* get(const edict_t* pPlayer)
+	{
+		return m_Bots[static_cast<size_t>(slotOfEdict(pPlayer))];
+	}
+
 	static int levelInit(); //TODO: Not implemented? [APG]RoboCop[CL]
 
 private:
